Handled edge case with identical input and output paths, implemented simpson rule

+++ fund1/libmath/libmath/math.c
@@ -21,7 +21,7 @@ long double combinations(const int n, const int k) {
   return factorial(n) / (factorial(k) * factorial(n - k));
 }
 
-long double combinations_incr_k(const int n, const int k, const int c_n_k) {
+long long int combinations_incr_k(const long double n, const long double k, const long double c_n_k) {
   return c_n_k * (n - k) / (k + 1);
 }
